requirements.txt
# Core ML
torch>=2.1.0
numpy>=1.24.0

# Data & tokenization
datasets>=2.16.0
tokenizers>=0.15.0
huggingface_hub>=0.20.0

# SVG processing
lxml>=4.9.0
cairosvg>=2.7.0

# Scaling law fitting
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
Pillow>=10.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
pandas>=2.0.0

# µP (Phase 3)
mup>=1.0.0
